printCharImgs/requirements.txt
numpy~=1.24.3
pillow~=10.1.0
PyMuPDF
freetype-py
fonttools~=4.46.0
keras~=2.13.1
pdfminer.six
Augmentor~=0.2.12
opencv-python~=4.8.1.78
scikit-image~=0.21.0
tensorflow~=2.13.0
split-folders
pdfreader
levenshtein~=0.24.0
requests~=2.31.0
